+++ src/components/Experiences/Experiences.tsx
@@ -1,13 +1,7 @@
 import { useEffect, useState } from "react"
 import type { Experience } from "../../types/Experience"
 import axios from "axios"
-import Card from "../Card/Card"
-import Loading from "../Loading/Loading"
-import Text from "../Text/Text"
-import SeeMore from "../../features/SeeMore/SeeMore"
-import { useDispatch, useSelector } from "react-redux"
-import { setSeeMoreClicked } from "../../features/SeeMore/SeeMoreSlice"
-import type { RootState } from "../../store"
+import Item from "../Item/Item"
 
 interface ExperiencesProps {
     personId?: number | null
@@ -20,7 +14,6 @@ const Experiences: React.FC<ExperiencesProps> = ({
     isLoading
 }) => {
     const [experiences, setExperiences] = useState<Experience[] | null>(null)
-    const [hoveredItemId, setHoveredItemId] = useState<number | null>(null)
 
     useEffect(() => {
         const  fetchData = async () => {
@@ -44,86 +37,8 @@ const Experiences: React.FC<ExperiencesProps> = ({
         fetchData();
     }, [personId])
 
-    const handleMouseEnter = (itemId:number) => {
-        setHoveredItemId(itemId)
-    }
-    const handleMouseLeave = () => {
-        setHoveredItemId(null)
-    }
-    
-    const SeeMoreState = useSelector((state: RootState) => state.SeeMore)
-    const dispatch = useDispatch();
-
-    if(!isLoading && experiences) {
-        return (
-            <Card title={componentId} boxShadow={true} isLoading={isLoading} onMouseLeave={() => handleMouseLeave()} content={
-                <div className="column sm-row-gap">
-                    {experiences?.map((item)=>(
-                        <Card key={item.id} isLoading={isLoading} boxShadow={true} onMouseEnter={() => handleMouseEnter(item.id)} type="card" content={
-                            <div className="column sm-row-gap">
-                                <div className="row flex sm-column-gap sm-row-gap">
-                                    <Text type="h3" isLoading={isLoading} content={item.name}></Text>
-                                    <span className="row flex sm-row-gap sm-column-gap">
-                                        <Card type="smCard" isLoading={isLoading} classes="bg-light-blue" title={item.position}></Card>
-                                        <Card type="smCard" isLoading={isLoading} classes={item.end_date ? "bg-orange" : "bg-green"} title={item.end_date ? item.end_date : "Empregado aqui"}></Card>
-                                    </span>
-                                </div>
-                                <div className="responsive-row">
-                                    <Card isLoading={isLoading} type="card" classes="bg-light-grey center" content={
-                                        <>
-                                            <div className="row sm-row-gap flex">
-                                                <div className="column" style={{flexGrow: "1"}}>
-                                                    <span>
-                                                        <Text type="h5" isLoading={isLoading} content={
-                                                            <>
-                                                                <strong>Empresa: </strong>{item.enterprise}
-                                                            </>
-                                                            }></Text>
-                                                    </span>
-                                                    <span>
-                                                        <Text type="h5" isLoading={isLoading} content={
-                                                            <>
-                                                                <strong>Localização: </strong>{item.address}
-                                                            </>
-                                                        }></Text>
-                                                    </span>
-                                                </div>
-                                                <div className="column" style={{flexGrow: "1"}}>
-                                                    <span>
-                                                        <Text type="h5" isLoading={isLoading} style={{flexGrow: "1"}} content={
-                                                            <>
-                                                                <strong>Início: </strong>{item.start_date}
-                                                            </>
-                                                        }></Text>
-                                                    </span>
-                                                    <span>
-                                                        <Text type="h5" isLoading={isLoading} style={{flexGrow: "1"}}content={
-                                                            <>
-                                                                <strong>Fim: </strong>{item.end_date ? item.end_date : "Até o momento"}
-                                                            </>
-                                                        }></Text>
-                                                    </span>
-                                                </div>
-                                                <div className={`${"expansive-container"} ${(SeeMoreState.componentId === componentId && SeeMoreState.seeMoreId === item.id) && "active"}`}>
-                                                    <p>{item.about}</p>
-                                                </div>
-                                            </div>
-                                        </>
-                                    }></Card>
-                                    <SeeMore componentId={componentId} seeMoreId={item.id} isActive={hoveredItemId === item.id} onClick={() => {
-                                        dispatch(setSeeMoreClicked({seeMoreId: item.id, componentId: componentId}))
-                                        }}></SeeMore>
-                                </div>
-                            </div>
-                        }></Card>
-                    ))}
-                </div>
-            }></Card>
-        )
-    } else {
-        return (
-            <Loading title={componentId}></Loading>
-        )
-    }
+    return (
+        <Item items={experiences} componentId={componentId} isLoading={isLoading}></Item>
+    )
 }
 export default Experiences
